requirements.txt
python-telegram-bot==21.9
python-dotenv
google-generativeai
groq
anthropic
psycopg2-binary
cryptography
httpx